+++ rollup.config.js
@@ -3,20 +3,27 @@ import external from 'rollup-plugin-peer-deps-external';
 import del from 'rollup-plugin-delete';
 import { terser } from 'rollup-plugin-terser';
 import pkg from './package.json';
+import typescript from 'rollup-plugin-typescript2';
 
-export default {
-  input: pkg.source,
-  output: [
-    { file: pkg.main, format: 'cjs' },
-    { file: pkg.module, format: 'esm' },
-  ],
-  plugins: [
-    external(),
-    terser(),
-    babel({
-      exclude: 'node_modules/**',
-    }),
-    del({ targets: ['dist/*'] }),
-  ],
-  external: Object.keys(pkg.peerDependencies || {}),
-};
+export default [
+  {
+    input: pkg.source,
+    output: [
+      { file: pkg.main, format: 'cjs', exports: 'named' },
+      { file: pkg.module, format: 'esm', exports: 'named' },
+    ],
+    plugins: [
+      del({ targets: ['dist/*'] }),
+      terser(),
+      external(),
+      babel({
+        exclude: 'node_modules/**',
+      }),
+      typescript({
+        rollupCommonJSResolveHack: false,
+        clean: true,
+      }),
+    ],
+    external: Object.keys(pkg.peerDependencies || {}),
+  },
+];